requirements.txt
matplotlib==3.8.3
mediapipe==0.10.11
numpy==1.24.3
opencv_python==4.9.0.80
openpyxl==3.1.2
optuna==3.6.1
pillow==10.2.0
pygame==2.5.2
pywin32==306
scipy==1.14.0
kivy==2.3.0
google==3.0.0
protobuf==3.20.3
torch==2.2.2+cu121